testing lettuce heads method on laplace

+++ Block_wise_GPS_correction.py
@@ -1488,7 +1488,7 @@ class Patch:
 		blue_channel = self.rgb_img[:,:,0].copy()
 
 		img = green_channel-0.61*blue_channel-0.39*red_channel
-
+		img =img.astype('uint8')
 		min_p = np.amin(img)
 		max_p = np.amax(img)
